+++ src/domain/mod.rs
@@ -24,6 +24,7 @@ pub mod rack_view;
 pub mod site_view;
 pub mod tag_view;
 pub mod tenant_view;
+pub mod util;
 pub mod vlan_view;
 pub mod vm_view;